+++ tests/check/check_test.go
@@ -48,9 +48,6 @@ func testRunTestMatrix(t *testing.T, engine string, experimental bool) {
 			goleak.VerifyNone(t)
 		})
 		cfg := config.MustDefaultConfig()
-		if experimental {
-			cfg.Experimentals = append(cfg.Experimentals, "enable-check-optimizations")
-		}
 		cfg.Log.Level = "error"
 		cfg.Datastore.Engine = engine
 		cfg.ListUsersDeadline = 0   // no deadline
@@ -327,7 +324,6 @@ func testRunAll(t *testing.T, engine string) {
 		goleak.VerifyNone(t)
 	})
 	cfg := config.MustDefaultConfig()
-	cfg.Experimentals = append(cfg.Experimentals, "enable-check-optimizations")
 	cfg.Log.Level = "error"
 	cfg.Datastore.Engine = engine
 
@@ -394,7 +390,6 @@ const githubModel = `
 // and a cancellation function that stops the benchmark timer.
 func setupBenchmarkTest(b *testing.B, engine string) (openfgav1.OpenFGAServiceClient, context.CancelFunc) {
 	cfg := config.MustDefaultConfig()
-	cfg.Experimentals = append(cfg.Experimentals, "enable-check-optimizations")
 	cfg.Log.Level = "none"
 	cfg.Datastore.Engine = engine